Consistent file{s,set,list} naming; bug fix.
Naming:
"files" means a list of absolute paths to files.
"fileset" means a list of tuples ({+,-}, abs path, rel path).
"filelist" is a list of (assumed relative) paths given by the user.
"stack" is a list of filesets.

The bug fix was that readStack was returning a list with the same
fileset possibly many times, and that fileset would include all
files in the entire .fhstack file, as opposed to splitting along
"===" lines.